+++ packages/dev/serializers/src/3MF/core/model/index.ts
@@ -4,6 +4,4 @@ export * from "./3mf.opc";
 export * from "./3mf";
 export * from "./3mf.builder";
 export * from "./3mf.types";
-export * from "./3mf.math";
 export * from "./3mf.serializer";
-export * from "./3mf.utils";
